Revert "Split components out into libraries for reuse with slang-vscode-extension (#157)"
This reverts commit fc06c7ede8e3d69f6170cf90003df30b21f9b7f6.
