--- engine/playground-render-canvas/src/index.ts
@@ -1,3 +1,0 @@
-import RenderCanvas from './RenderCanvas.vue';
-
-export { RenderCanvas };
