--- engine/playground-render-canvas/tsconfig.json
@@ -1,9 +1,0 @@
-{
-	"compilerOptions": {
-		"rootDir": "src",
-	},
-	"include": [
-		"src"
-	],
-	"extends": "../tsconfig.base.json"
-}
