--- engine/playground-render-canvas/vite.config.ts
@@ -1,24 +1,0 @@
-import { defineConfig } from 'vite';
-import vue from '@vitejs/plugin-vue';
-import dts from 'vite-plugin-dts';
-import path from 'path';
-
-export default defineConfig({
-    plugins: [vue(), dts()],
-    build: {
-        lib: {
-            entry: path.resolve(__dirname, 'src/index.ts'),
-            name: 'RenderCanvasLib',
-            fileName: (format) => `index.${format}.js`,
-        },
-        rollupOptions: {
-            external: ['vue'],
-            output: {
-                globals: {
-                    vue: 'Vue',
-                },
-            },
-        },
-        sourcemap: true,
-    }
-});
