--- engine/playground-uniform-panel/src/index.ts
@@ -1,3 +1,0 @@
-import UniformPanel from './UniformPanel.vue';
-
-export { UniformPanel };
