--- engine/shared/tsconfig.json
@@ -1,17 +1,0 @@
-{
-	"extends": "../tsconfig.base.json",
-	"compilerOptions": {
-		"lib": [
-			"es2020",
-			"WebWorker",
-			"DOM"
-		],
-		"rootDir": "src",
-		"composite": true,
-		"outDir": "dist"
-	},
-	"include": [
-		"src"
-	],
-	"exclude": []
-}
