--- engine/slang-compilation-engine/media/.gitignore
@@ -1,2 +1,0 @@
-slang-wasm.*
-spirv-tools.*
