--- engine/slang-compilation-engine/src/index.ts
@@ -1,3 +1,0 @@
-export * from './compiler.js';
-export * from './compilationUtils.js';
-export * from './playgroundCompiler.js';
