--- engine/slang-compilation-engine/src/slang/loader.d.ts
@@ -1,4 +1,0 @@
-declare module '*.slang?raw' {
-	const content: string;
-	export default content;
-}
