--- engine/slang-compilation-engine/tsconfig.json
@@ -1,12 +1,0 @@
-{
-    "extends": "../tsconfig.base.json",
-    "compilerOptions": {
-        "outDir": "dist",
-        "rootDir": "src",
-        "strict": true,
-        "declaration": true
-    },
-    "include": [
-        "src"
-    ]
-}
