--- engine/tsconfig.base.json
@@ -1,10 +1,0 @@
-{
-    "compilerOptions": {
-        "module": "esnext",
-        "target": "esnext",
-        "lib": ["esnext", "WebWorker", "DOM"],
-        "sourceMap": true,
-        "types": ["@webgpu/types"],
-        "moduleResolution": "bundler",
-    }
-}
