+++ package.json
@@ -3,13 +3,9 @@
   "version": "0.0.0",
   "private": true,
   "type": "module",
-  "workspaces": [
-    "engine/*"
-  ],
   "scripts": {
     "dev": "vite",
-    "build": "npm run build-engine && run-p type-check build-only",
-    "build-engine": "npm run build -w slang-playground-shared && npm run build --workspaces --exclude=slang-playground-shared",
+    "build": "run-p type-check \"build-only {@}\" --",
     "preview": "vite preview",
     "build-only": "vite build",
     "type-check": "vue-tsc --build"
@@ -22,12 +18,8 @@
     "lucide-vue-next": "^0.473.0",
     "monaco-editor": "^0.52.0",
     "pako": "^2.1.0",
-    "playground-render-canvas": "0.0.1",
-    "playground-uniform-panel": "0.0.1",
-    "slang-compilation-engine": "0.0.1",
-    "slang-playground-shared": "0.0.1",
     "splitpanes": "^3.1.8",
-    "vue": "^3.5.17"
+    "vue": "^3.5.13"
   },
   "devDependencies": {
     "@tsconfig/node22": "^22.0.0",
@@ -40,7 +32,7 @@
     "@vue/tsconfig": "^0.7.0",
     "@webgpu/types": "^0.1.51",
     "npm-run-all2": "^7.0.2",
-    "typescript": "^5.8.3",
+    "typescript": "~5.6.3",
     "vite": "^6.0.5",
     "vite-plugin-vue-devtools": "^7.6.8",
     "vue-tsc": "^2.1.10"
