+++ src/components/ui/Colorpick.vue
@@ -40,7 +40,7 @@ function updateColor(event: Event) {
     flex-direction: row;
     align-items: center;
     min-width: 0;
-    color: var(--vscode-editor-foreground, white);
+    color: white;
 
 }
