+++ src/language-server.ts
@@ -1,4 +1,5 @@
 import { slangd } from './try-slang.js';
+import playgroundSource from "./slang/playground.slang?raw";
 import type { CompletionContext } from './slang-wasm.js';
 
 import * as monaco from "monaco-editor";
@@ -8,7 +9,6 @@ import jsonWorker from 'monaco-editor/esm/vs/language/json/json.worker?worker';
 import cssWorker from 'monaco-editor/esm/vs/language/css/css.worker?worker';
 import htmlWorker from 'monaco-editor/esm/vs/language/html/html.worker?worker';
 import tsWorker from 'monaco-editor/esm/vs/language/typescript/ts.worker?worker';
-import { PLAYGROUND_SOURCE } from 'slang-compilation-engine';
 
 export const userCodeURI = "file:///user.slang";
 const playgroundCodeURI = "file:///playground.slang";
@@ -858,7 +858,7 @@ export function initLanguageServer() {
         throw new Error("Slang is undefined!");
     }
     slangd.didOpenTextDocument(userCodeURI, "");
-    slangd.didOpenTextDocument(playgroundCodeURI, PLAYGROUND_SOURCE);
+    slangd.didOpenTextDocument(playgroundCodeURI, playgroundSource);
 }
 
 export function translateSeverity(severity: number) {
