+++ src/pass_through.ts
@@ -1,4 +1,5 @@
 export const passThroughshaderCode = `
+
       struct VertexShaderOutput {
         @builtin(position) position: vec4f,
         @location(0) texcoord: vec2f,
@@ -13,7 +14,7 @@ export const passThroughshaderCode = `
           vec2f( 1.0,  0.0),  // right, center
           vec2f( 0.0,  1.0),  // center, top
 
-          // 2nd triangle
+          // 2st triangle
           vec2f( 0.0,  1.0),  // center, top
           vec2f( 1.0,  0.0),  // right, center
           vec2f( 1.0,  1.0),  // right, top
@@ -25,7 +26,7 @@ export const passThroughshaderCode = `
           vec2f( 1.0,  -1.0),  // right, center
           vec2f( -1.0,  1.0),  // center, top
 
-          // 2nd triangle
+          // 2st triangle
           vec2f( -1.0,  1.0),  // center, top
           vec2f( 1.0,  -1.0),  // right, center
           vec2f( 1.0,  1.0),  // right, top
@@ -85,12 +86,10 @@ export class GraphicsPipeline {
             vertex:
             {
                 module: shaderModule,
-                entryPoint: "vs",
             },
             fragment:
             {
                 module: shaderModule,
-                entryPoint: "fs",
                 targets: [{ format: navigator.gpu.getPreferredCanvasFormat(), }],
             },
         });
