+++ src/try-slang.ts
@@ -1,5 +1,4 @@
-import { PLAYGROUND_SOURCE, SlangCompiler } from 'slang-compilation-engine';
-import { RUNNABLE_ENTRY_POINT_NAMES } from "slang-playground-shared";
+import { SlangCompiler, RUNNABLE_ENTRY_POINT_NAMES } from './compiler.js';
 import { fetchWithProgress } from './util.js';
 import type { LanguageServer, MainModule } from "./slang-wasm.js";
 import type { PublicApi as JJsonPublicApi } from "jjsontree.js/src/ts/api.js"
@@ -65,19 +64,15 @@ createModule(moduleConfig).then((module) => {
     if (label)
         label.innerText = "Initializing Slang Compiler...";
     try {
-        {
-            let FS = module.FS;
-            FS.writeFile("/playground.slang", PLAYGROUND_SOURCE);
-        }
         compiler = new SlangCompiler(module);
         let result = compiler.init();
         slangd = module.createLanguageServer();
-        if (result.succ) {
+        if (result.ret) {
             moduleLoadingMessage = "Slang compiler initialized successfully.\n";
             window.dispatchEvent(new CustomEvent('slangLoaded', {}));
         }
         else {
-            console.log(result.message);
+            console.log(result.msg);
             moduleLoadingMessage = "Failed to initialize Slang Compiler.\n";
             if (label)
                 label.innerText = moduleLoadingMessage;
