+++ tsconfig.app.json
@@ -3,13 +3,21 @@
   "include": [
     "env.d.ts",
     "src/**/*",
-    "src/**/*.vue"
+    "src/**/*.vue",
+    "compiler.ts",
+    "compute.ts",
+    "demos/demo-list.ts",
+    "language-server.ts",
+    "src/main.ts",
+    "pass_through.ts",
+    "playgroundShader.ts",
+    "try-slang.ts",
+    "util.ts"
 , "public/jsontree"  ],
   "exclude": [
     "src/**/__tests__/*"
   ],
   "compilerOptions": {
-    "baseUrl": ".",
     "tsBuildInfoFile": "./node_modules/.tmp/tsconfig.app.tsbuildinfo",
     "paths": {
       "@/*": [
